In allbills, detect when bills change title.
The unfortunate NM practice is to file a bunch of "dummy bills"
just before the filing deadline, with names like "PUBLIC PEACE,
HEALTH, SAFETY & WELFARE", then change them later to get around
the filing deadline. So the billtracker needs to flag those as
new bills in the allbills page when their title changes, so
users can be aware of them.

This, combined with unit tests that try to test that, turned out
to require a lot of broad-ranging changes. On the bright side,
this also includes a commit that vastly speeds up the allbills page
(at the cost of not being able to show number of users tracking
each bill, but you can use the Popular Bills page for that, and
new bills won't normally have anyone tracking them anyway.