Better handling of House/Senate "third reading".
Add a time (hardwired to 11 am, since the floor schedules don't list
their meeting time) ) for bills listed on the House or Senate's
"third reading" list.

Highlight (in yellow) bills that are scheduled, but drop the highlight
for bills that are on the House or Senate floors but not listed as
being in the "third reading".

Let api/all_committees be called with a GET, to make it easier to test.