More logging when creating and confirming users.
I'm seeing occasional failures in the log, but can't tell whether
they're failures confirming legit users, or attacks trying to
break in.  Also made the username more obvious in the confirmation
email, and bumped up the number and variety of characters in the
reset string in case these really are breakin attempts I'm seeing.