Various changes to make things clearer for first-timers.
Always update last_check time when showing the home page
(previously it wasn't updated on a user's first check, so it
said "this is your first check" for longer than it should have).
If the home page is showing no bills, add some helpful text
explaining the various ways of finding bills to add.
Move things around a bit in the password reset, and log more
detail in case users report problems.