Rewrite captcha as a singleton, and make tests pass.
The captcha is now a singleton, and the captcha object is no
longer stored in the session; this potentially exposed answers
as well as questions in cookies if anyone cared to look.
Now only the question (capq) is stored in the session, and
it is passed in to the singleton ChattyCaptcha object when
validating answers.

Meanwhile, Flask has drastically changed what tests need to do
to set up things like the db, request context, and the session,
requiring a significant rewrite of unit tests.
test_billtracker.py seems to work right now;
fingers crossed that it keeps working.