Huge changes to support allbills*.json.
This introduces a cached JSON file (which will eventually be moved
into the database) that stores history of bill name changes.
This fleshes out the information in the allbills page, and also
means that refreshing allbills will pull in FIR, LESC and amendment
updates, so those no longer need to be called.