Detect committee subs as amendments.
It turns out that dummy bills, even if they don't change their
title or original text, get committee subs in place of bill text.
These are in the same directory as the original bill text, but with
a different filename, so the link to the committee sub needs to be
shown on the allbills page.