Try to work around alembic/sqlalchemy problems.
The last migration failed on the server, and after I hacked the
db by hand, for a while I got a lot of obscure alembic errors.
Some of these changes are definitely necessary, others may not be.
Also, the two new fields were too small, and it turns out postgres
actually pays attention to that and errors if you set a field to
a string that's too long, so I made them bigger.